Complete attestation_validation, bar tests
