+++ beacon_chain/attestation_validation/src/lib.rs
@@ -7,12 +7,17 @@ extern crate types;
 
 #[macro_use]
 mod macros;
+
+mod block_inclusion;
 mod enums;
-mod validate_for_block;
-mod validate_for_state;
-mod validate_signature;
+mod justified_block;
+mod justified_slot;
+mod shard_block;
+mod signature;
 
 pub use enums::{Invalid, Outcome, Error};
-pub use validate_for_block::validate_attestation_for_block;
-pub use validate_for_state::validate_attestation_data_for_state;
-pub use validate_signature::validate_attestation_signature;
+pub use block_inclusion::validate_attestation_for_block;
+pub use justified_slot::validate_attestation_justified_slot;
+pub use justified_block::validate_attestation_justified_block_hash;
+pub use signature::validate_attestation_signature;
+pub use shard_block::validate_attestation_data_shard_block_hash;
