+++ beacon_chain/types/src/attestation_data.rs
@@ -12,7 +12,7 @@ pub const SSZ_ATTESTION_DATA_LENGTH: usize = {
     32 // justified_block_hash
 };
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Default)]
 pub struct AttestationData {
     pub slot: u64,
     pub shard: u64,
